+++ backend/UserService/UserService.Api/Controllers/UserController.cs
@@ -1,13 +1,9 @@
-﻿using System.Net.Http.Headers;
-using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.AspNetCore.Authorization;
+﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Authorization;
-using Microsoft.Extensions.Options;
-using UserService.Api.Filters;
 using UserService.Api.Interfaces;
 using UserService.Application.Dto;
-using UserService.DataAccess.Handlers.Jwt;
+using UserService.Application.Dto.EmailDtos;
+using UserService.DataAccess.Enums;
 
 namespace UserService.Api.Controllers;
 
@@ -15,9 +11,13 @@ namespace UserService.Api.Controllers;
 [Route("users")]
 [Authorize]
 public class UserController(
-    IUserService service
+    IUserService service,
+    IAuthenticationService authService
     ) : Controller
 {
+    /// <summary>
+    /// Get all users with pagination
+    /// </summary>
     [HttpGet]
     public async Task<IResult> GetUsers([FromQuery] PaginatedPageUsers query, CancellationToken cancellationToken)
     {
@@ -25,6 +25,9 @@ public class UserController(
         return Results.Ok(resultUsers);
     }
     
+    /// <summary>
+    /// Get user by ID
+    /// </summary>
     [HttpGet("{id:Guid}")]
     public async Task<IResult> GetUser([FromRoute] Guid id, CancellationToken cancellationToken)
     {
@@ -32,28 +35,76 @@ public class UserController(
         return Results.Ok(resultUsers);
     }
     
-    [HttpPut("add-role")]
-    [Authorize(Policy = "Admin")]
-    public async Task<IResult> AddAdminRoleToUser([FromQuery] AddRoleDto addRoleDto, CancellationToken cancellationToken)
+    /// <summary>
+    /// Register a new user
+    /// </summary>
+    [HttpPost]
+    public async Task<IResult> Register(RegisterDto user, CancellationToken cancellationToken)
     {
-        var resultUser = await service.AddRole(addRoleDto, cancellationToken);
+        var resultUser = await authService.Register(user, cancellationToken);
+        
         return Results.Ok(resultUser);
     }
     
-    [HttpPut("update")]
+    /// <summary>
+    /// Update user information
+    /// </summary>
+    [HttpPut("{id:Guid}")]
     [Authorize]
-    public async Task<IResult> UpdateUser(UpdateUserDto user, CancellationToken cancellationToken)
+    public async Task<IResult> UpdateUser([FromRoute] Guid id, UpdateUserDto user, CancellationToken cancellationToken)
     {
-        var resultUser = await service.UpdateUser(user, cancellationToken);
+        var resultUser = await service.UpdateUser(id, user, cancellationToken);
         return Results.Ok(resultUser);
     }
 
-    [HttpPost("soft-delete")]
+    /// <summary>
+    /// Soft delete user account
+    /// </summary>
+    [HttpDelete]
     [Authorize]
     public async Task<IResult> SoftDeleteUser(CancellationToken cancellationToken)
     {
         var accessToken = HttpContext.Request.Headers.Authorization.ToString().Replace("Bearer ", string.Empty);
         var deletedUser = await service.SoftDelete(accessToken, cancellationToken);
         return Results.Ok(deletedUser);
+    }
+    
+    /// <summary>
+    /// Initiate account restoration process
+    /// </summary>
+    [HttpPost("{email}/restore")]
+    public async Task<IResult> RestoreAccount([FromRoute] string email, CancellationToken cancellationToken)
+    {
+        var callbackUrl = Url.RouteUrl(
+            "RestoreAccount",
+            values: null,
+            protocol: Request.Scheme);
+    
+        var token = await authService.RecoverAccountAsync(email, callbackUrl!, cancellationToken);
+    
+        return Results.Ok(token);
+    }
+    
+    /// <summary>
+    /// Complete account restoration
+    /// </summary>
+    [HttpGet("restore", Name = "RestoreAccount")]
+    public async Task<IResult> OnRestoreAccount([FromQuery] EmailTokenDto restoreAccountRequest, CancellationToken cancellationToken)
+    {
+        var result = await authService.RestoreAccountAsync(restoreAccountRequest, cancellationToken);
+    
+        return Results.Ok(result);
+    }
+    
+    /// <summary>
+    /// Add role to user (Admin only)
+    /// </summary>
+    [HttpPut("{id:Guid}/roles")]
+    [Authorize(Policy = "Admin")]
+    public async Task<IResult> AddAdminRoleToUser([FromRoute]Guid id, [FromQuery] Roles role, CancellationToken cancellationToken)
+    {
+        var dto = new AddRoleDto{Role = role, UserId = id};
+        var resultUser = await service.AddRole(dto, cancellationToken);
+        return Results.Ok(resultUser);
     }
 }
