+++ backend/UserService/UserService.Application/Dto/UserDtos/UpdateUserDto.cs
@@ -2,7 +2,6 @@
 
 public class UpdateUserDto
 {
-    public Guid Id { get; set; }
     public string Username { get; set; } = string.Empty;
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
@@ -12,9 +11,8 @@ public class UpdateUserDto
         
     }
 
-    public UpdateUserDto(Guid id, string username, string firstName, string lastName)
+    public UpdateUserDto(string username, string firstName, string lastName)
     {
-        Id = id;
         Username = username;
         FirstName = firstName;
         LastName = lastName;
